feat: XYNE-55280 revert-chunking-for-some-doc-type
Services-Requiring-Redeployment:
  - backend
